Move installer scripts to separate file
